This fixes a bug where image files were permanently deleted while disassociating them from a file. Now it checks to see if the picture is used anywhere else, and only permanently deletes them if nothing references it anymore.

+++ web/pluto-admin/operations/mediaBrowser/editMediaFile.php
@@ -316,17 +316,25 @@ function editMediaFile($output,$mediadbADO,$dbADO) {
 			$toDelete=$_REQUEST['picID'];
 			$path=stripslashes(@$_REQUEST['path']);
 			
-			$deletePic='DELETE FROM Picture WHERE PK_Picture=?';
-			$mediadbADO->Execute($deletePic,$toDelete);
-			
-			$deletePicAttribute='DELETE FROM Picture_Attribute WHERE FK_Picture=?';
-			$mediadbADO->Execute($deletePicAttribute,$toDelete);
-			
+			//remove reference to the picture
 			$deletePicFile='DELETE FROM Picture_File WHERE FK_Picture=?';
 			$mediadbADO->Execute($deletePicFile,$toDelete);
 			
-			@unlink($GLOBALS['mediaPicsPath'].$toDelete.'.jpg');
-			@unlink($GLOBALS['mediaPicsPath'].$toDelete.'_tn.jpg');
+			//Is anything else referencing this picture?
+			$pictureReference =         'SELECT * FROM Picture_Attribute WHERE FK_Picture = '.$toDelete;
+			$pictureReference .= ' UNION SELECT * FROM Picture_Disc WHERE FK_Picture = '.$toDelete;
+			$pictureReference .= ' UNION SELECT * FROM Picture_Download WHERE FK_Picture = '.$toDelete;
+			$pictureReference .= ' UNION SELECT * FROM Picture_File WHERE FK_Picture = '.$toDelete;
+			$numPictureReferences = $mediadbADO->Execute($pictureReference);
+
+			//Delete picture from database and file system only if nothing else is referencing it!
+			if($numPictureReferences->RecordCount()==0) {
+				$deletePic='DELETE FROM Picture WHERE PK_Picture=?';
+				$mediadbADO->Execute($deletePic,$toDelete);
+				@unlink($GLOBALS['mediaPicsPath'].$toDelete.'.jpg');
+				@unlink($GLOBALS['mediaPicsPath'].$toDelete.'_tn.jpg');
+				
+			}
 
 			$cmd='sudo -u root /usr/pluto/bin/UpdateMedia -w -d "'.bash_escape($path).'"';
 			exec_batch_command($cmd);
